move document router to routerlicious-base
