+++ server/routerlicious/packages/routerlicious-base/package.json
@@ -51,6 +51,7 @@
     "@fluidframework/protocol-definitions": "^0.1025.0",
     "@fluidframework/server-kafka-orderer": "^0.1025.0",
     "@fluidframework/server-lambdas": "^0.1025.0",
+    "@fluidframework/server-lambdas-driver": "^0.1025.0",
     "@fluidframework/server-memory-orderer": "^0.1025.0",
     "@fluidframework/server-services": "^0.1025.0",
     "@fluidframework/server-services-client": "^0.1025.0",
