+++ server/routerlicious/packages/routerlicious-base/src/utils/documentRouter.ts
