+++ server/routerlicious/packages/routerlicious-base/src/utils/index.ts
@@ -4,6 +4,7 @@
  */
 
 export * from "./constants";
+export * from "./documentRouter";
 export * from "./middleware";
 export * from "./params";
 export * from "./runner";
