+++ server/routerlicious/packages/routerlicious/src/deli/index.ts
@@ -4,14 +4,13 @@
  */
 
 import { BroadcasterLambda, DeliLambdaFactory } from "@fluidframework/server-lambdas";
+import { createDocumentRouter } from "@fluidframework/server-routerlicious-base";
 import { LocalKafka, LocalContext, LocalLambdaController } from "@fluidframework/server-memory-orderer";
 import * as services from "@fluidframework/server-services";
 import * as core from "@fluidframework/server-services-core";
 import { Provider } from "nconf";
 import { RedisOptions } from "ioredis";
 import * as winston from "winston";
-// eslint-disable-next-line import/no-internal-modules
-import { createDocumentRouter } from "../kafka-service/documentRouter";
 
 export async function deliCreate(config: Provider): Promise<core.IPartitionLambdaFactory> {
     const mongoUrl = config.get("mongo:endpoint") as string;
