+++ server/routerlicious/packages/routerlicious/src/event-hub/deli/index.ts
@@ -4,6 +4,7 @@
  */
 
 import { BroadcasterLambda, DeliLambdaFactory } from "@fluidframework/server-lambdas";
+import { createDocumentRouter } from "@fluidframework/server-routerlicious-base";
 import { LocalKafka, LocalContext, LocalLambdaController } from "@fluidframework/server-memory-orderer";
 import * as services from "@fluidframework/server-services";
 import * as core from "@fluidframework/server-services-core";
@@ -11,8 +12,6 @@ import { EventHubProducer } from "@fluidframework/server-services-ordering-event
 import { Provider } from "nconf";
 import { RedisOptions } from "ioredis";
 import * as winston from "winston";
-// eslint-disable-next-line import/no-internal-modules
-import { createDocumentRouter } from "../../kafka-service/documentRouter";
 
 export async function deliCreate(config: Provider): Promise<core.IPartitionLambdaFactory> {
     const mongoUrl = config.get("mongo:endpoint") as string;
