--- server/routerlicious/packages/routerlicious/src/test/tsconfig.json
@@ -1,22 +1,0 @@
-{
-    "extends": "@fluidframework/build-common/ts-common-config.json",
-    "compilerOptions": {
-        "strictNullChecks": false,
-        "rootDir": "./",
-        "outDir": "../../dist/test",
-        "types": [
-            "node",
-            "mocha"
-        ],
-        "declaration": false,
-        "declarationMap": false
-    },
-    "include": [
-        "./**/*"
-    ],
-    "references": [
-        {
-            "path": "../.."
-        }
-    ]
-}
